// src/components/stack/header/ios/StackHeaderItem.ios.types.ts
import type { ReactElement } from '@lynx-js/react';
import type { PlatformIconIOS } from '../../../shared/types.js';
import type { StackHeaderMenuIOS } from './StackHeaderMenu.ios.types.js';

export type StackHeaderItemPlacement =
  | 'leading'
  | 'trailing'
  | 'title'
  | 'subtitle'
  | 'largeSubtitle';

export type StackHeaderItemProps = {
  placement: StackHeaderItemPlacement;
  itemId?: string | undefined;
  title?: string | undefined;
  icon?: PlatformIconIOS | undefined;
  render?: (() => ReactElement) | undefined;
  menu?: StackHeaderMenuIOS | undefined;
  onPress?: (() => void) | undefined;
};